Give UntilStepButton a proper stacktrace
