+++ osu.Framework/Testing/Drawables/Steps/AssertButton.cs
@@ -4,7 +4,6 @@
 using System;
 using System.Diagnostics;
 using System.Text;
-using NUnit.Framework;
 using osuTK.Graphics;
 
 namespace osu.Framework.Testing.Drawables.Steps
@@ -44,18 +43,5 @@ namespace osu.Framework.Testing.Drawables.Steps
         }
 
         public override string ToString() => "Assert: " + base.ToString();
-
-        private class TracedException : AssertionException
-        {
-            private readonly StackTrace trace;
-
-            public TracedException(string description, StackTrace trace)
-                : base(description)
-            {
-                this.trace = trace;
-            }
-
-            public override string StackTrace => trace.ToString();
-        }
     }
 }
